Added dummy interrupt routine
Fixes an annoying bug in which the interrupt is triggered immediately
upon attachment. This should ultimately be fixed by clearing the ISR
flags, but I don't know how to do that in a clean, board-agnostic manner
since the Due allows you to attach the ISR to any digital pin.

Note, this commit is grafted on from the orphaned serial-usb branch in
AMOArduino. It was commit # 088e608ed